load questions one at a time and from multiple files
Allow questions to be stored in multiple .txt files in a directory.
Each time a new question is selected, a random file is opened and a
random line is selected from that file. This is suboptimal,  but works.

probability of any given question being chosen: (1/files)(1/lines),
thus files with fewer lines are penalized...